requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
anthropic==0.34.2
pyxirr==0.10.6
pandas==2.2.2
numpy==1.26.4
pdfplumber==0.11.4
python-dotenv==1.0.1
